+++ pyctr/type/save/common.py
@@ -13,9 +13,9 @@ from ..base import TypeReaderCryptoBase
 from .partition import Partition, load_partdesc
 
 if TYPE_CHECKING:
-    from os import PathLike
-    from typing import BinaryIO, Dict, Literal, Optional, Union
+    from typing import Dict, Literal, Optional
 
+    from ...common import FilePath, FilePathOrObject
     from ...crypto import CryptoEngine
     from .cmac import CMACTypeBase
 
@@ -58,10 +58,9 @@ class PartitionContainerBase(TypeReaderCryptoBase):
     _header: bytes
     """Raw header for CMAC generation."""
 
-    def __init__(self, file: 'Union[PathLike, str, bytes, BinaryIO]', mode: 'ReadWriteBinaryFileModes' = 'rb', *,
+    def __init__(self, file: 'FilePathOrObject', mode: 'ReadWriteBinaryFileModes' = 'rb', *,
                  closefd: 'Optional[bool]' = None, crypto: 'CryptoEngine' = None, dev: bool = False,
-                 cmac_base: 'CMACTypeBase' = None, sd_key_file: 'Union[PathLike, str, bytes]' = None,
-                 sd_key: bytes = None):
+                 cmac_base: 'CMACTypeBase' = None, sd_key_file: 'FilePath' = None, sd_key: bytes = None):
         super().__init__(file, closefd=closefd, mode=mode, crypto=crypto, dev=dev)
 
         self.cmac = self._file.read(0x10)