+++ pyctr/type/sd.py
@@ -18,6 +18,7 @@ from .sdtitle import SDTitleReader
 if TYPE_CHECKING:
     from os import PathLike
     from typing import BinaryIO, List, Union
+    from ..common import FilePath
 
     # noinspection PyProtectedMember
     from ..crypto import CTRFileIO
@@ -63,8 +64,8 @@ class SDFilesystem:
 
     __slots__ = ('_base_path', '_crypto', '_id0_path', 'current_id1', 'id1s')
 
-    def __init__(self, path: 'Union[PathLike, str, bytes]', *, crypto: CryptoEngine = None, dev: bool = False,
-                 sd_key_file: 'Union[PathLike, str, bytes]' = None, sd_key: bytes = None):
+    def __init__(self, path: 'FilePath', *, crypto: CryptoEngine = None, dev: bool = False,
+                 sd_key_file: 'FilePath' = None, sd_key: bytes = None):
         if crypto:
             self._crypto = crypto
         else:
@@ -200,6 +201,8 @@ class SDFilesystem:
         title_id = title_id.lower()
         sd_path = f'/title/{title_id[0:8]}/{title_id[8:16]}/content'
 
+        # not sure why PyCharm thinks this variable is unused?
+        # noinspection PyUnusedLocal
         tmd_path = None
         for f in self.listdir(sd_path):
             if f.endswith('.tmd'):