+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Some performance improvements were done to `CTRFileIO` and `RomFSReader`.
 * Add `nand` module with `NAND` class, to read and write to a NAND image
 * Add `__slots__` to a bunch of classes (`CCIReader`, `CDNReader`, `CIAReader`, `ExeFSReader`, `NAND`, `NCCHReader`, `RomFSReader`, `SDFilesystem`, `SDTitleReader`, `SMDH`, `ConfigSaveReader`, `TypeReaderBase`, `TypeReaderCryptoBase`)
 * Update copyright year
+* Various documentation and type hint changes
+  * Add new `FilePath` and `FilePathOrObject` types
 
 ## v0.6.0 - January 26, 2022
 ### Highlights